% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{AwesomeOscillatorHistogram}
\alias{AwesomeOscillatorHistogram}
\title{AwesomeOscillatorHistogram indicator}
\usage{
AwesomeOscillatorHistogram(fast, slow, sma_period)
}
\description{
AwesomeOscillatorHistogram indicator
}
\keyword{internal}
